% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{albedo}
\alias{albedo}
\title{Calculates Broadband Albedo from Landsat data}
\usage{
albedo(path = getwd(), image.SR, aoi, coeff = "Tasumi", sat = "auto",
  ESPA = FALSE, result.folder = NULL)
}
\description{
Calculates Broadband Albedo from Landsat data
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

