% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.Clustering.R
\name{benchmark}
\alias{benchmark}
\title{Benchmarking
This function plots a group of radar charts to benchmark a customer against his/her cluster or to compare the different clusters
The comparison/benchmarking is done in terms of consumption averages for three different modalities (daily, weekly, seasonlly)
For this function to work, two steps must be completed beforehand:
1) a cluster assignment to be calculated using calculate_clusters function
2) consumption aggregates to be calculated using calculate_consumption_aggregates}
\usage{
benchmark(clust_obj, aggregates_list, service_id = NULL,
  plot_title = "")
}
\arguments{
\item{clust_obj}{: a cluster object returned from calculate_clusters function}

\item{aggregates_list}{: a list of aggregate consumptions calculated using calculate_consumption_aggregates function}

\item{service_id}{: ID of the customer to be benchmarked against his/her cluster, if left None, all clusters will be plotted}

\item{plot_title}{: desired title of the generated plot}
}
\description{
Benchmarking
This function plots a group of radar charts to benchmark a customer against his/her cluster or to compare the different clusters
The comparison/benchmarking is done in terms of consumption averages for three different modalities (daily, weekly, seasonlly)
For this function to work, two steps must be completed beforehand:
1) a cluster assignment to be calculated using calculate_clusters function
2) consumption aggregates to be calculated using calculate_consumption_aggregates
}
