% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{calc.SR}
\alias{calc.SR}
\title{Calculates surface reflectance for L7}
\usage{
calc.SR(path = getwd(), image.TOAr, sat = "auto", ESPA = FALSE,
  format = "tif", aoi, result.folder = NULL, incidence.hor, WeatherStation,
  surface.model)
}
\description{
Calculates surface reflectance for L7
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

