% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{calc.TOAr}
\alias{calc.TOAr}
\title{Calculates Top of atmosphere reflectance}
\usage{
calc.TOAr(path = getwd(), image.DN, sat = "auto", ESPA = FALSE, aoi,
  result.folder = NULL, incidence.rel)
}
\description{
Calculates Top of atmosphere reflectance
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
LPSO. (2004). Landsat 7 science data users handbook, Landsat Project Science Office, NASA Goddard Space Flight Center, Greenbelt, Md., (http://ltpwww.gsfc.nasa.gov/IAS/handbook/handbook_toc.html) (Feb. 5, 2007)
}

