% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_sensibleHeatFlux.R
\name{calcAnchors}
\alias{calcAnchors}
\title{Select anchors pixels for H function}
\usage{
calcAnchors(image, Ts, LAI, albedo, Z.om, n = 1,
  anchors.method = "CITRA-MCB", sat = "auto", ESPA = F, plots = TRUE,
  deltaTemp = 5, verbose = FALSE, ...)
}
\description{
Select anchors pixels for H function
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
CITRA y MCB (com pers)
}

