% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_sensibleHeatFlux.R
\name{calcH2}
\alias{calcH2}
\title{Iterative function to estimate H and R.ah}
\usage{
calcH2(anchors, Ts, Z.om, WeatherStation, ETp.coef = 1.05, Z.om.ws = 0.0018,
  sat = "auto", ESPA = F, mountainous = FALSE, DEM, Rn, G,
  verbose = FALSE)
}
\description{
Iterative function to estimate H and R.ah
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

