% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{calcTOAr}
\alias{calcTOAr}
\title{Calculates Top of atmosphere reflectance}
\usage{
calcTOAr(path = getwd(), image.DN, sat = "auto", ESPA = FALSE, aoi,
  incidence.rel, MTL, ...)
}
\description{
Calculates Top of atmosphere reflectance
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
LPSO. (2004). Landsat 7 science data users handbook, Landsat Project Science Office, NASA Goddard Space Flight Center, Greenbelt, Md., (http://landsathandbook.gsfc.nasa.gov/) (Feb. 5, 2007)
}

