% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.Clustering.R
\name{calculate_clusters}
\alias{calculate_clusters}
\title{calculate_clusters
This function receives a features dataframe (custom as per YVW requirements or obtained from one of the feature extraction functions in this package)
Based on the features provided and the maximum allowed number of clusters, the function will calculate the best cluster assignment possible and will
in an iterative manner while removing any outliers found in each iteration. Finally, the function will return the cluster assignment and some other
useful plots. The output of this function can be used as an input for the customer_benchmarking function which benchmarks a customer based on his/her
cluster.}
\usage{
calculate_clusters(features, k_override = NA, k_max = 10)
}
\arguments{
\item{features}{: is a dataframe with service_id as the first column and the features as the remaining columns}

\item{k_override}{: can be set to an integer number to override the number of clusters, if k is left NULL, best K will be automatically chosen based on lowest DB index}

\item{k_max}{: is the maximum number of clusters allowed}
}
\value{
: return a list containing the chosen number of clusters (k_chosen), the cluster centers scaled and unscaled, the cluster_assignment data frame, and two plot objects
for a PCA representation, a cluster barplot representation, and a vector of outlier service IDs.
}
\description{
calculate_clusters
This function receives a features dataframe (custom as per YVW requirements or obtained from one of the feature extraction functions in this package)
Based on the features provided and the maximum allowed number of clusters, the function will calculate the best cluster assignment possible and will
in an iterative manner while removing any outliers found in each iteration. Finally, the function will return the cluster assignment and some other
useful plots. The output of this function can be used as an input for the customer_benchmarking function which benchmarks a customer based on his/her
cluster.
}
