% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.Clustering.R
\name{calculate_consumption_aggregates}
\alias{calculate_consumption_aggregates}
\title{calculate_consumption_aggregates
Receives a timeseries list and calculates consumption aggregate values for three modalities (daily, weekly, seasonally)
The output can be saved to disk to be used at a later time for benchmarking using radar charts.}
\usage{
calculate_consumption_aggregates(timeseries_list, filename = NA)
}
\arguments{
\item{timeseries_list}{: a list of xts timeseries objects containing customer meter data}

\item{filename}{: optional argument with the filename to save to disk, if left NA, the aggregates won't be saved}
}
\value{
the function returns a list of three dataframes: daily_aggregate, weekly_aggregate, and sesonal_aggregate
}
\description{
calculate_consumption_aggregates
Receives a timeseries list and calculates consumption aggregate values for three modalities (daily, weekly, seasonally)
The output can be saved to disk to be used at a later time for benchmarking using radar charts.
}
