% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{checkSRTMgrids}
\alias{checkSRTMgrids}
\title{Check needed SRTM grids from image extent}
\usage{
checkSRTMgrids(raw.image, path = getwd(), format = "tif")
}
\description{
Check needed SRTM grids from image extent
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}

