% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicsStyle.R
\docType{data}
\name{colBlack}
\alias{colBlack}
\title{Graphics Style Constants for the water package}
\format{
An object of class \code{character} of length 1.
}
\usage{
colBlack
}
\description{
These colour and icon definitions are used across all graphical outputs.
Colours follow a themed scheme around water, safety, and severity.
}
\keyword{datasets}
