% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.Visualisation.R
\name{consumption_cal}
\alias{consumption_cal}
\title{consumption_cal}
\usage{
consumption_cal(df, mID = "", year = NA)
}
\arguments{
\item{df}{xts object of the meter consumption}

\item{mID}{meterID to be used in the plot title}

\item{year}{the year to be plotted. default NA which plots all years available for the contract}
}
\value{
calendar plot
}
\description{
Plots  consumption as a calendar plot. This can help to see the patterns in consumption at multiple granularity levels
}
\examples{
\dontrun{ consumption_cal(xts, mID='1234', year=2016)}
}
