% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.Visualisation.R
\name{consumption_cal_HM}
\alias{consumption_cal_HM}
\title{consumption_cal_HM}
\usage{
consumption_cal_HM(df, mID = "", year = NA, scale = "continuous",
  fixed_scale = TRUE)
}
\arguments{
\item{df}{xts object of the meter consumption}

\item{mID}{meterID to be used in the plot title}

\item{year}{the year to be plotted. default NA which plots all years available for the contract}

\item{scale}{type of scale used for the heatmap, continues or discrete}

\item{fixed_scale}{Boolean, to indicate whether the scale should be fixed for all the plots or change dyanmically per plot. We recommend keeping the scale fixed to enable comparison over different contracts. Default is TRUE}
}
\value{
consumption heatmap over calendar plot
}
\description{
plots the consumption as a heat map over calendar plot.
}
\examples{
\dontrun{consumption_cal_HM(contract1, mID='1254', year=2016, scale= 'continuous',fixed_scale=TRUE )}
}
