% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.Visualisation.R
\name{consumption_day}
\alias{consumption_day}
\title{consumption_day}
\usage{
consumption_day(df, mID = "", year = NA)
}
\arguments{
\item{df}{xts object of the meter consumption}

\item{mID}{meterID to be used in the plot title}

\item{year}{the year to be plotted. default NA which plots all years available for the contract}
}
\value{
faceted line chart plot
}
\description{
return faceted line chart of consumption per hour of the day for every weekday on summer and winter
}
\examples{
\dontrun{consumption_day(rds$'11', mID='11', year=2014)}
}
