% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.Visualisation.R
\name{consumption_heat}
\alias{consumption_heat}
\title{consumption_heat}
\usage{
consumption_heat(df, mID = "", year = NA, scale = "continuous",
  fixed_scale = TRUE)
}
\arguments{
\item{df}{xts object of the meter consumption}

\item{mID}{meterID to be used in the plot title}

\item{year}{the year to be plotted. default NA which plots all years available for the contract}

\item{scale}{type of scale used for the heatmap, continues or discrete}

\item{fixed_scale}{Boolean, to indicate whether the scale should be fixed for all the plots or change dyanmically per plot. We recommend keeping the scale fixed to enable comparison over different contracts. Default is TRUE}
}
\value{
Consumption heat plot
}
\description{
THis plot generates consumption heat tile, with date on the x-axis, hour of the day on the y-axis and consumption depicted as coloured tile.
}
\examples{
\dontrun{Consumption_Heat(rds$`1_101510`, mID= " ",year=NA,scale = 'discrete',fixed_scale = TRUE)}
}
