% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.Visualisation.R
\name{consumption_vis}
\alias{consumption_vis}
\title{consumption_vis}
\usage{
consumption_vis(df, contracts_list = NA, scale = "discrete",
  fixed_scale = TRUE, file_name = "p", path = getwd(),
  output_format = c("png", "pdf"))
}
\arguments{
\item{df}{xts object of the meter consumption}

\item{contracts_list}{list of contracts to be investigated}

\item{scale}{type of scale used for the heatmap, continues or discrete}

\item{fixed_scale}{Boolean, to indicate whether the scale should be fixed for all the plots or change dyanmically per plot. We recommend keeping the scale fixed to enable comparison over different contracts. Default is TRUE.}

\item{file_name}{name of the file to be used for outputing the plotting results}

\item{path}{name of the path where plots will be outputed}

\item{output_format}{png or pdf format}
}
\value{
pdf of png output with the results
}
\description{
Generate diagnostic plots to analyse the consumption patterns
}
\examples{
\dontrun{consumption_vis(contract_lst, contract_list= NA, scale= 'discrete', fixed_scale= TRUE)}
}
