% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{create.aoi}
\alias{create.aoi}
\title{Create aoi from topleft and bottomright coordinates}
\usage{
create.aoi(topleft = c(x, y), bottomright = c(x, y))
}
\arguments{
\item{topleft}{a vector with topleft coordinates}

\item{bottomright}{a vector with bottomright coordinates}
}
\value{
object of class SpatialPolygons-class
}
\description{
Create aoi from topleft and bottomright coordinates
}

