% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Forecasting.R
\name{customer_forecast_ets}
\alias{customer_forecast_ets}
\title{customer_forecast_ets}
\usage{
customer_forecast_ets(object, plot_type = c("bar", "line"),
  pdf_file = NA, out_csv = TRUE, mID = "000")
}
\arguments{
\item{object}{Contract timeseries object or list of contracts timeseries}

\item{plot_type}{Type of plot used to dispaly the forecast results, line or bar charts}

\item{pdf_file}{name of pdf file to output the pltos. If kept empty no pdf will be generated}

\item{out_csv}{boolean to indicate  if forecast need to be outputed in csv files}

\item{mID}{meter ID for the case wheere a single contract is passed, used to plot the contract id in the plot title}
}
\value{
list of forecast results
}
\description{
Forecast cutomers monthly consumption for the next 6 months. This function uses ETS model and require at least 2 years of consumption histor
to generate forecast.
}
\examples{
\dontrun{customer_forecast_ets(contractss, plot_type= 'bar', pdf_file= 'forecast.pdf', out_csv = TRUE)}
}
