% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.Utilities.R, R/Leak_Detection_MIN.R
\name{daily_min}
\alias{daily_min}
\title{daily_min}
\usage{
daily_min(input)

daily_min(input)
}
\arguments{
\item{input}{: contract timeseries object or list of contracts timeseries}

\item{rds}{list of xts objects}
}
\value{
list of xts objects with daily minimum

list or Single timeseries object of daily minimum consumption
}
\description{
Utility function to get the minimum daily ocnsumption per contract

Utility function to get the minimum daily ocnsumption per contract
}
\examples{
\dontrun{daily_min(contract)}
}
