% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_ET.R
\name{dailyEToPM}
\alias{dailyEToPM}
\title{Calculates daily ET using Penman Monteith hourly formula for every hour}
\usage{
dailyEToPM(WeatherStation, DOY, long.z = WeatherStation$long)
}
\arguments{
\item{WeatherStation}{a data frame with all the needed fields (see example)}

\item{DOY}{day of year}

\item{long.z}{longitude for local time}
}
\value{
ET daily in mm.h-1
}
\description{
Calculates daily ET using Penman Monteith hourly formula for every hour
}
\author{
Guillermo F Olmedo
}
\references{
Allen 2005 ASCE
}

