% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.Utilities.R
\name{data_prep}
\alias{data_prep}
\title{Data Preparation}
\usage{
data_prep(rds, align_interval = NULL, nas.action = c("interpolate",
  "LOCF", "None"))
}
\arguments{
\item{rds}{a list of xts object returned from read.csv function}

\item{align_interval}{the timing between subsequent consumption measurments in minutes}

\item{nas.action}{flag to decide on the action with NA points: Interpolate conducts linear interpoaltion from last observation
LOCF Last observation carried forward
None: leave NAs as they are}
}
\value{
Standerdizedd list of temseries per input  contract
}
\description{
This function takes rds with a list of contracts and returns them in standerdized equidistant form
}
\examples{
\dontrun{data_prep(contracts, align_interval = 30, nas.action = LOCF)}
}
