% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{ET.24h}
\alias{ET.24h}
\title{Calculates ET-24hs from energy balance and Weather Station}
\usage{
ET.24h(Rn, G, H, Ts, WeatherStation, ETr.daily, C.rad = 1)
}
\description{
Calculates ET-24hs from energy balance and Weather Station
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

