% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02.Clustering.R
\name{extract_features}
\alias{extract_features}
\title{extract_features
This function extract Time Series features from a list of time series objects. The features extracted are based on
tsfeatuers package and suppport multiple seasonalities. The purpose of the feature extraction is to use the features
(or part thereof) for clustering. Please see the example markdown for a complete workflow.}
\usage{
extract_features(timeseries_list)
}
\arguments{
\item{timeseries_list}{: a list of timeseries XTS objects named with the service_id of each customer contract.}
}
\value{
: a features dataframe where the first column is the service_id and the remaining columns are the features
(refer to documentation for more details about the features descritpion)
}
\description{
extract_features
This function extract Time Series features from a list of time series objects. The features extracted are based on
tsfeatuers package and suppport multiple seasonalities. The purpose of the feature extraction is to use the features
(or part thereof) for clustering. Please see the example markdown for a complete workflow.
}
