% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_WeatherStation.R
\name{getDataWS}
\alias{getDataWS}
\title{Export data.frame from waterWeatherStation Object}
\usage{
getDataWS(WeatherStation)
}
\description{
Export data.frame from waterWeatherStation Object
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}

