% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions_rah.iteractive_CITRA.R
\name{H.CITRA}
\alias{H.CITRA}
\title{Iterative function to estimate H and R.ah - CITRA method}
\usage{
H.CITRA(anchors, image, Ts, LAI, albedo, Z.om, n = 1,
  anchors.method = "CITRA-MCB", WeatherStation, ETp.coef = 1.05,
  Z.om.ws = 0.0018, sat = "auto", ESPA = F, mountainous = FALSE, DEM,
  Rn, G, plots = TRUE, deltaTemp = 5, result.folder = NULL)
}
\description{
Iterative function to estimate H and R.ah - CITRA method
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
CITRA y MCB (com pers)
}

