% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.Forecasting.R
\name{hourly_forecasst}
\alias{hourly_forecasst}
\title{hourly_forecasst}
\usage{
hourly_forecasst(object, plot_type = c("bar", "line"), temp_file,
  pdf_file = NA, out_csv = TRUE)
}
\arguments{
\item{object}{Contract timeseries object or list of contracts timeseries}

\item{plot_type}{Type of plot used to dispaly the forecast results, line or bar charts}

\item{pdf_file}{name of pdf file to output the pltos. If kept empty no pdf will be generated}

\item{out_csv}{boolean to indicate  if forecast need to be outputed in csv files}
}
\value{
forecasted series
}
\description{
Conduct 24 hour consumption forecast for aggregatre consumption from the list of contracts passed to the function
}
\examples{
\dontrun{hourly_forecast(contractss, plot_type= 'line', pdf_file= 'forecast.pdf', out_csv = TRUE)}
}
