% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_ET.R
\name{hourlyET}
\alias{hourlyET}
\title{Calculates ET using Penman Monteith hourly formula}
\usage{
hourlyET(WeatherStation, hours, DOY, long.z = WeatherStation$long)
}
\arguments{
\item{WeatherStation}{a data frame with all the needed fields (see example)}

\item{hours}{time of the day in hours in 24hs format}

\item{DOY}{day of year}

\item{long.z}{longitude for local time}
}
\value{
ET hourly in mm.h-1
}
\description{
Calculates ET using Penman Monteith hourly formula
}
\examples{
WeatherStation  <- data.frame(wind=4.72,
                              RH=59,
                              temp=24.3,
                              radiation=675,
                              height=2.2,
                              lat=-35.37,
                              long=71.5946,
                              elev=124)
  hourlyET(WeatherStation, hours=10.5, DOY=363, long.z=71.635)
}
\author{
Guillermo F Olmedo
}
\references{
Allen 2005 ASCE
}

