% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{incLWradiation}
\alias{incLWradiation}
\title{Calculates long wave incoming radiation}
\usage{
incLWradiation(WeatherStation, DEM, solar.angles, Ts)
}
\arguments{
\item{WeatherStation}{Weather Station data}

\item{DEM}{digital elevation model in meters.}

\item{solar.angles}{rasterStack with latitude, declination, hour.angle, incidence.hor and incidence.rel. See solarAngles()}

\item{Ts}{Land surface temperature. See surfaceTemperature()}
}
\description{
Calculates long wave incoming radiation
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

