% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{incoming.lw.radiation}
\alias{incoming.lw.radiation}
\title{Calculates long wave incoming radiation}
\usage{
incoming.lw.radiation(WeatherStation, DEM, solar.angles, result.folder = NULL)
}
\description{
Calculates long wave incoming radiation
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

