% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L7_Talca}
\alias{L7_Talca}
\title{Landsat 7 scene from east of Talca, Chile}
\format{\preformatted{Formal class 'RasterStack' [package "raster"] with 11 slots
  ..@ filename: chr ""
  ..@ layers  :List of 7
  .. ..$ :Formal class 'RasterLayer' [package "raster"] with 12 slots
  .. .. .. ..@ file    :Formal class '.RasterFile' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ name        : chr ""
  .. .. .. .. .. ..@ datanotation: chr "FLT4S"
  .. .. .. .. .. ..@ byteorder   : chr "little"
  .. .. .. .. .. ..@ nodatavalue : num -32768
  .. .. .. .. .. ..@ NAchanged   : logi TRUE
  .. .. .. .. .. ..@ nbands      : int 1
  .. .. .. .. .. ..@ bandorder   : chr "BIL"
  .. .. .. .. .. ..@ offset      : int 0
  .. .. .. .. .. ..@ toptobottom : logi TRUE
  .. .. .. .. .. ..@ blockrows   : int 0
  .. .. .. .. .. ..@ blockcols   : int 0
  .. .. .. .. .. ..@ driver      : chr ""
  .. .. .. .. .. ..@ open        : logi FALSE
  .. .. .. ..@ data    :Formal class '.SingleLayerData' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ values    : int [1:202686] 50 48 50 48 46 51 52 54 54 51 ...
  .. .. .. .. .. ..@ offset    : num 0
  .. .. .. .. .. ..@ gain      : num 1
  .. .. .. .. .. ..@ inmemory  : logi TRUE
  .. .. .. .. .. ..@ fromdisk  : logi FALSE
  .. .. .. .. .. ..@ isfactor  : logi FALSE
  .. .. .. .. .. ..@ attributes: list()
  .. .. .. .. .. ..@ haveminmax: logi TRUE
  .. .. .. .. .. ..@ min       : int 1
  .. .. .. .. .. ..@ max       : int 255
  .. .. .. .. .. ..@ band      : int 1
  .. .. .. .. .. ..@ unit      : chr ""
  .. .. .. .. .. ..@ names     : chr "B"
  .. .. .. ..@ legend  :Formal class '.RasterLegend' [package "raster"] with 5 slots
  .. .. .. .. .. ..@ type      : chr(0) 
  .. .. .. .. .. ..@ values    : logi(0) 
  .. .. .. .. .. ..@ color     : logi(0) 
  .. .. .. .. .. ..@ names     : logi(0) 
  .. .. .. .. .. ..@ colortable: logi(0) 
  .. .. .. ..@ title   : chr(0) 
  .. .. .. ..@ extent  :Formal class 'Extent' [package "raster"] with 4 slots
  .. .. .. .. .. ..@ xmin: num 273105
  .. .. .. .. .. ..@ xmax: num 288045
  .. .. .. .. .. ..@ ymin: num -3926655
  .. .. .. .. .. ..@ ymax: num -3914445
  .. .. .. ..@ rotated : logi FALSE
  .. .. .. ..@ rotation:Formal class '.Rotation' [package "raster"] with 2 slots
  .. .. .. .. .. ..@ geotrans: num(0) 
  .. .. .. .. .. ..@ transfun:function ()  
  .. .. .. ..@ ncols   : int 498
  .. .. .. ..@ nrows   : int 407
  .. .. .. ..@ crs     :Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. .. ..@ projargs: chr "+proj=utm +zone=19 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
  .. .. .. ..@ history : list()
  .. .. .. ..@ z       : list()
  .. ..$ :Formal class 'RasterLayer' [package "raster"] with 12 slots
  .. .. .. ..@ file    :Formal class '.RasterFile' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ name        : chr ""
  .. .. .. .. .. ..@ datanotation: chr "FLT4S"
  .. .. .. .. .. ..@ byteorder   : chr "little"
  .. .. .. .. .. ..@ nodatavalue : num -32768
  .. .. .. .. .. ..@ NAchanged   : logi TRUE
  .. .. .. .. .. ..@ nbands      : int 1
  .. .. .. .. .. ..@ bandorder   : chr "BIL"
  .. .. .. .. .. ..@ offset      : int 0
  .. .. .. .. .. ..@ toptobottom : logi TRUE
  .. .. .. .. .. ..@ blockrows   : int 0
  .. .. .. .. .. ..@ blockcols   : int 0
  .. .. .. .. .. ..@ driver      : chr ""
  .. .. .. .. .. ..@ open        : logi FALSE
  .. .. .. ..@ data    :Formal class '.SingleLayerData' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ values    : int [1:202686] 42 44 43 39 39 43 46 48 47 44 ...
  .. .. .. .. .. ..@ offset    : num 0
  .. .. .. .. .. ..@ gain      : num 1
  .. .. .. .. .. ..@ inmemory  : logi TRUE
  .. .. .. .. .. ..@ fromdisk  : logi FALSE
  .. .. .. .. .. ..@ isfactor  : logi FALSE
  .. .. .. .. .. ..@ attributes: list()
  .. .. .. .. .. ..@ haveminmax: logi TRUE
  .. .. .. .. .. ..@ min       : int 19
  .. .. .. .. .. ..@ max       : int 189
  .. .. .. .. .. ..@ band      : int 1
  .. .. .. .. .. ..@ unit      : chr ""
  .. .. .. .. .. ..@ names     : chr "G"
  .. .. .. ..@ legend  :Formal class '.RasterLegend' [package "raster"] with 5 slots
  .. .. .. .. .. ..@ type      : chr(0) 
  .. .. .. .. .. ..@ values    : logi(0) 
  .. .. .. .. .. ..@ color     : logi(0) 
  .. .. .. .. .. ..@ names     : logi(0) 
  .. .. .. .. .. ..@ colortable: logi(0) 
  .. .. .. ..@ title   : chr(0) 
  .. .. .. ..@ extent  :Formal class 'Extent' [package "raster"] with 4 slots
  .. .. .. .. .. ..@ xmin: num 273105
  .. .. .. .. .. ..@ xmax: num 288045
  .. .. .. .. .. ..@ ymin: num -3926655
  .. .. .. .. .. ..@ ymax: num -3914445
  .. .. .. ..@ rotated : logi FALSE
  .. .. .. ..@ rotation:Formal class '.Rotation' [package "raster"] with 2 slots
  .. .. .. .. .. ..@ geotrans: num(0) 
  .. .. .. .. .. ..@ transfun:function ()  
  .. .. .. ..@ ncols   : int 498
  .. .. .. ..@ nrows   : int 407
  .. .. .. ..@ crs     :Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. .. ..@ projargs: chr "+proj=utm +zone=19 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
  .. .. .. ..@ history : list()
  .. .. .. ..@ z       : list()
  .. ..$ :Formal class 'RasterLayer' [package "raster"] with 12 slots
  .. .. .. ..@ file    :Formal class '.RasterFile' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ name        : chr ""
  .. .. .. .. .. ..@ datanotation: chr "FLT4S"
  .. .. .. .. .. ..@ byteorder   : chr "little"
  .. .. .. .. .. ..@ nodatavalue : num -32768
  .. .. .. .. .. ..@ NAchanged   : logi TRUE
  .. .. .. .. .. ..@ nbands      : int 1
  .. .. .. .. .. ..@ bandorder   : chr "BIL"
  .. .. .. .. .. ..@ offset      : int 0
  .. .. .. .. .. ..@ toptobottom : logi TRUE
  .. .. .. .. .. ..@ blockrows   : int 0
  .. .. .. .. .. ..@ blockcols   : int 0
  .. .. .. .. .. ..@ driver      : chr ""
  .. .. .. .. .. ..@ open        : logi FALSE
  .. .. .. ..@ data    :Formal class '.SingleLayerData' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ values    : int [1:202686] 44 48 50 42 39 49 56 56 56 53 ...
  .. .. .. .. .. ..@ offset    : num 0
  .. .. .. .. .. ..@ gain      : num 1
  .. .. .. .. .. ..@ inmemory  : logi TRUE
  .. .. .. .. .. ..@ fromdisk  : logi FALSE
  .. .. .. .. .. ..@ isfactor  : logi FALSE
  .. .. .. .. .. ..@ attributes: list()
  .. .. .. .. .. ..@ haveminmax: logi TRUE
  .. .. .. .. .. ..@ min       : int 16
  .. .. .. .. .. ..@ max       : int 120
  .. .. .. .. .. ..@ band      : int 1
  .. .. .. .. .. ..@ unit      : chr ""
  .. .. .. .. .. ..@ names     : chr "R"
  .. .. .. ..@ legend  :Formal class '.RasterLegend' [package "raster"] with 5 slots
  .. .. .. .. .. ..@ type      : chr(0) 
  .. .. .. .. .. ..@ values    : logi(0) 
  .. .. .. .. .. ..@ color     : logi(0) 
  .. .. .. .. .. ..@ names     : logi(0) 
  .. .. .. .. .. ..@ colortable: logi(0) 
  .. .. .. ..@ title   : chr(0) 
  .. .. .. ..@ extent  :Formal class 'Extent' [package "raster"] with 4 slots
  .. .. .. .. .. ..@ xmin: num 273105
  .. .. .. .. .. ..@ xmax: num 288045
  .. .. .. .. .. ..@ ymin: num -3926655
  .. .. .. .. .. ..@ ymax: num -3914445
  .. .. .. ..@ rotated : logi FALSE
  .. .. .. ..@ rotation:Formal class '.Rotation' [package "raster"] with 2 slots
  .. .. .. .. .. ..@ geotrans: num(0) 
  .. .. .. .. .. ..@ transfun:function ()  
  .. .. .. ..@ ncols   : int 498
  .. .. .. ..@ nrows   : int 407
  .. .. .. ..@ crs     :Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. .. ..@ projargs: chr "+proj=utm +zone=19 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
  .. .. .. ..@ history : list()
  .. .. .. ..@ z       : list()
  .. ..$ :Formal class 'RasterLayer' [package "raster"] with 12 slots
  .. .. .. ..@ file    :Formal class '.RasterFile' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ name        : chr ""
  .. .. .. .. .. ..@ datanotation: chr "FLT4S"
  .. .. .. .. .. ..@ byteorder   : chr "little"
  .. .. .. .. .. ..@ nodatavalue : num -32768
  .. .. .. .. .. ..@ NAchanged   : logi TRUE
  .. .. .. .. .. ..@ nbands      : int 1
  .. .. .. .. .. ..@ bandorder   : chr "BIL"
  .. .. .. .. .. ..@ offset      : int 0
  .. .. .. .. .. ..@ toptobottom : logi TRUE
  .. .. .. .. .. ..@ blockrows   : int 0
  .. .. .. .. .. ..@ blockcols   : int 0
  .. .. .. .. .. ..@ driver      : chr ""
  .. .. .. .. .. ..@ open        : logi FALSE
  .. .. .. ..@ data    :Formal class '.SingleLayerData' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ values    : int [1:202686] 69 69 61 0 0 0 0 0 0 70 ...
  .. .. .. .. .. ..@ offset    : num 0
  .. .. .. .. .. ..@ gain      : num 1
  .. .. .. .. .. ..@ inmemory  : logi TRUE
  .. .. .. .. .. ..@ fromdisk  : logi FALSE
  .. .. .. .. .. ..@ isfactor  : logi FALSE
  .. .. .. .. .. ..@ attributes: list()
  .. .. .. .. .. ..@ haveminmax: logi TRUE
  .. .. .. .. .. ..@ min       : int 0
  .. .. .. .. .. ..@ max       : int 146
  .. .. .. .. .. ..@ band      : int 1
  .. .. .. .. .. ..@ unit      : chr ""
  .. .. .. .. .. ..@ names     : chr "NIR"
  .. .. .. ..@ legend  :Formal class '.RasterLegend' [package "raster"] with 5 slots
  .. .. .. .. .. ..@ type      : chr(0) 
  .. .. .. .. .. ..@ values    : logi(0) 
  .. .. .. .. .. ..@ color     : logi(0) 
  .. .. .. .. .. ..@ names     : logi(0) 
  .. .. .. .. .. ..@ colortable: logi(0) 
  .. .. .. ..@ title   : chr(0) 
  .. .. .. ..@ extent  :Formal class 'Extent' [package "raster"] with 4 slots
  .. .. .. .. .. ..@ xmin: num 273105
  .. .. .. .. .. ..@ xmax: num 288045
  .. .. .. .. .. ..@ ymin: num -3926655
  .. .. .. .. .. ..@ ymax: num -3914445
  .. .. .. ..@ rotated : logi FALSE
  .. .. .. ..@ rotation:Formal class '.Rotation' [package "raster"] with 2 slots
  .. .. .. .. .. ..@ geotrans: num(0) 
  .. .. .. .. .. ..@ transfun:function ()  
  .. .. .. ..@ ncols   : int 498
  .. .. .. ..@ nrows   : int 407
  .. .. .. ..@ crs     :Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. .. ..@ projargs: chr "+proj=utm +zone=19 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
  .. .. .. ..@ history : list()
  .. .. .. ..@ z       : list()
  .. ..$ :Formal class 'RasterLayer' [package "raster"] with 12 slots
  .. .. .. ..@ file    :Formal class '.RasterFile' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ name        : chr ""
  .. .. .. .. .. ..@ datanotation: chr "FLT4S"
  .. .. .. .. .. ..@ byteorder   : chr "little"
  .. .. .. .. .. ..@ nodatavalue : num -32768
  .. .. .. .. .. ..@ NAchanged   : logi TRUE
  .. .. .. .. .. ..@ nbands      : int 1
  .. .. .. .. .. ..@ bandorder   : chr "BIL"
  .. .. .. .. .. ..@ offset      : int 0
  .. .. .. .. .. ..@ toptobottom : logi TRUE
  .. .. .. .. .. ..@ blockrows   : int 0
  .. .. .. .. .. ..@ blockcols   : int 0
  .. .. .. .. .. ..@ driver      : chr ""
  .. .. .. .. .. ..@ open        : logi FALSE
  .. .. .. ..@ data    :Formal class '.SingleLayerData' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ values    : int [1:202686] 0 0 0 0 0 0 0 0 0 0 ...
  .. .. .. .. .. ..@ offset    : num 0
  .. .. .. .. .. ..@ gain      : num 1
  .. .. .. .. .. ..@ inmemory  : logi TRUE
  .. .. .. .. .. ..@ fromdisk  : logi FALSE
  .. .. .. .. .. ..@ isfactor  : logi FALSE
  .. .. .. .. .. ..@ attributes: list()
  .. .. .. .. .. ..@ haveminmax: logi TRUE
  .. .. .. .. .. ..@ min       : int 0
  .. .. .. .. .. ..@ max       : int 138
  .. .. .. .. .. ..@ band      : int 1
  .. .. .. .. .. ..@ unit      : chr ""
  .. .. .. .. .. ..@ names     : chr "SWIR1"
  .. .. .. ..@ legend  :Formal class '.RasterLegend' [package "raster"] with 5 slots
  .. .. .. .. .. ..@ type      : chr(0) 
  .. .. .. .. .. ..@ values    : logi(0) 
  .. .. .. .. .. ..@ color     : logi(0) 
  .. .. .. .. .. ..@ names     : logi(0) 
  .. .. .. .. .. ..@ colortable: logi(0) 
  .. .. .. ..@ title   : chr(0) 
  .. .. .. ..@ extent  :Formal class 'Extent' [package "raster"] with 4 slots
  .. .. .. .. .. ..@ xmin: num 273105
  .. .. .. .. .. ..@ xmax: num 288045
  .. .. .. .. .. ..@ ymin: num -3926655
  .. .. .. .. .. ..@ ymax: num -3914445
  .. .. .. ..@ rotated : logi FALSE
  .. .. .. ..@ rotation:Formal class '.Rotation' [package "raster"] with 2 slots
  .. .. .. .. .. ..@ geotrans: num(0) 
  .. .. .. .. .. ..@ transfun:function ()  
  .. .. .. ..@ ncols   : int 498
  .. .. .. ..@ nrows   : int 407
  .. .. .. ..@ crs     :Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. .. ..@ projargs: chr "+proj=utm +zone=19 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
  .. .. .. ..@ history : list()
  .. .. .. ..@ z       : list()
  .. ..$ :Formal class 'RasterLayer' [package "raster"] with 12 slots
  .. .. .. ..@ file    :Formal class '.RasterFile' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ name        : chr ""
  .. .. .. .. .. ..@ datanotation: chr "INT2S"
  .. .. .. .. .. ..@ byteorder   : chr "little"
  .. .. .. .. .. ..@ nodatavalue : num -32768
  .. .. .. .. .. ..@ NAchanged   : logi FALSE
  .. .. .. .. .. ..@ nbands      : int 1
  .. .. .. .. .. ..@ bandorder   : chr "BIL"
  .. .. .. .. .. ..@ offset      : int 0
  .. .. .. .. .. ..@ toptobottom : logi TRUE
  .. .. .. .. .. ..@ blockrows   : int 1
  .. .. .. .. .. ..@ blockcols   : int 8081
  .. .. .. .. .. ..@ driver      : chr ""
  .. .. .. .. .. ..@ open        : logi FALSE
  .. .. .. ..@ data    :Formal class '.SingleLayerData' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ values    : int [1:202686] 0 0 0 0 0 0 0 0 0 0 ...
  .. .. .. .. .. ..@ offset    : num 0
  .. .. .. .. .. ..@ gain      : num 1
  .. .. .. .. .. ..@ inmemory  : logi TRUE
  .. .. .. .. .. ..@ fromdisk  : logi FALSE
  .. .. .. .. .. ..@ isfactor  : logi FALSE
  .. .. .. .. .. ..@ attributes: list()
  .. .. .. .. .. ..@ haveminmax: logi TRUE
  .. .. .. .. .. ..@ min       : int 0
  .. .. .. .. .. ..@ max       : int 163
  .. .. .. .. .. ..@ band      : int 1
  .. .. .. .. .. ..@ unit      : chr ""
  .. .. .. .. .. ..@ names     : chr "thermal.low"
  .. .. .. ..@ legend  :Formal class '.RasterLegend' [package "raster"] with 5 slots
  .. .. .. .. .. ..@ type      : chr(0) 
  .. .. .. .. .. ..@ values    : logi(0) 
  .. .. .. .. .. ..@ color     : logi(0) 
  .. .. .. .. .. ..@ names     : logi(0) 
  .. .. .. .. .. ..@ colortable: logi(0) 
  .. .. .. ..@ title   : chr(0) 
  .. .. .. ..@ extent  :Formal class 'Extent' [package "raster"] with 4 slots
  .. .. .. .. .. ..@ xmin: num 273105
  .. .. .. .. .. ..@ xmax: num 288045
  .. .. .. .. .. ..@ ymin: num -3926655
  .. .. .. .. .. ..@ ymax: num -3914445
  .. .. .. ..@ rotated : logi FALSE
  .. .. .. ..@ rotation:Formal class '.Rotation' [package "raster"] with 2 slots
  .. .. .. .. .. ..@ geotrans: num(0) 
  .. .. .. .. .. ..@ transfun:function ()  
  .. .. .. ..@ ncols   : int 498
  .. .. .. ..@ nrows   : int 407
  .. .. .. ..@ crs     :Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. .. ..@ projargs: chr "+proj=utm +zone=19 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
  .. .. .. ..@ history : list()
  .. .. .. ..@ z       : list()
  .. ..$ :Formal class 'RasterLayer' [package "raster"] with 12 slots
  .. .. .. ..@ file    :Formal class '.RasterFile' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ name        : chr ""
  .. .. .. .. .. ..@ datanotation: chr "FLT4S"
  .. .. .. .. .. ..@ byteorder   : chr "little"
  .. .. .. .. .. ..@ nodatavalue : num -32768
  .. .. .. .. .. ..@ NAchanged   : logi TRUE
  .. .. .. .. .. ..@ nbands      : int 1
  .. .. .. .. .. ..@ bandorder   : chr "BIL"
  .. .. .. .. .. ..@ offset      : int 0
  .. .. .. .. .. ..@ toptobottom : logi TRUE
  .. .. .. .. .. ..@ blockrows   : int 0
  .. .. .. .. .. ..@ blockcols   : int 0
  .. .. .. .. .. ..@ driver      : chr ""
  .. .. .. .. .. ..@ open        : logi FALSE
  .. .. .. ..@ data    :Formal class '.SingleLayerData' [package "raster"] with 13 slots
  .. .. .. .. .. ..@ values    : int [1:202686] 0 0 0 0 0 0 0 0 0 0 ...
  .. .. .. .. .. ..@ offset    : num 0
  .. .. .. .. .. ..@ gain      : num 1
  .. .. .. .. .. ..@ inmemory  : logi TRUE
  .. .. .. .. .. ..@ fromdisk  : logi FALSE
  .. .. .. .. .. ..@ isfactor  : logi FALSE
  .. .. .. .. .. ..@ attributes: list()
  .. .. .. .. .. ..@ haveminmax: logi TRUE
  .. .. .. .. .. ..@ min       : int 0
  .. .. .. .. .. ..@ max       : int 185
  .. .. .. .. .. ..@ band      : int 1
  .. .. .. .. .. ..@ unit      : chr ""
  .. .. .. .. .. ..@ names     : chr "SWIR2"
  .. .. .. ..@ legend  :Formal class '.RasterLegend' [package "raster"] with 5 slots
  .. .. .. .. .. ..@ type      : chr(0) 
  .. .. .. .. .. ..@ values    : logi(0) 
  .. .. .. .. .. ..@ color     : logi(0) 
  .. .. .. .. .. ..@ names     : logi(0) 
  .. .. .. .. .. ..@ colortable: logi(0) 
  .. .. .. ..@ title   : chr(0) 
  .. .. .. ..@ extent  :Formal class 'Extent' [package "raster"] with 4 slots
  .. .. .. .. .. ..@ xmin: num 273105
  .. .. .. .. .. ..@ xmax: num 288045
  .. .. .. .. .. ..@ ymin: num -3926655
  .. .. .. .. .. ..@ ymax: num -3914445
  .. .. .. ..@ rotated : logi FALSE
  .. .. .. ..@ rotation:Formal class '.Rotation' [package "raster"] with 2 slots
  .. .. .. .. .. ..@ geotrans: num(0) 
  .. .. .. .. .. ..@ transfun:function ()  
  .. .. .. ..@ ncols   : int 498
  .. .. .. ..@ nrows   : int 407
  .. .. .. ..@ crs     :Formal class 'CRS' [package "sp"] with 1 slot
  .. .. .. .. .. ..@ projargs: chr "+proj=utm +zone=19 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
  .. .. .. ..@ history : list()
  .. .. .. ..@ z       : list()
  ..@ title   : chr(0) 
  ..@ extent  :Formal class 'Extent' [package "raster"] with 4 slots
  .. .. ..@ xmin: num 273105
  .. .. ..@ xmax: num 288045
  .. .. ..@ ymin: num -3926655
  .. .. ..@ ymax: num -3914445
  ..@ rotated : logi FALSE
  ..@ rotation:Formal class '.Rotation' [package "raster"] with 2 slots
  .. .. ..@ geotrans: num(0) 
  .. .. ..@ transfun:function ()  
  ..@ ncols   : int 498
  ..@ nrows   : int 407
  ..@ crs     :Formal class 'CRS' [package "sp"] with 1 slot
  .. .. ..@ projargs: chr "+proj=utm +zone=19 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
  ..@ history : list()
  ..@ z       : list()
}}
\source{
\url{http://www.usgs.gov/}
}
\usage{
L7_Talca
}
\description{
A RasterStack object with a Landsat 7 ETM+ subset image.
Band names are stored in layer names. Images are USGS L1 raw data.
}
\details{
Metadata for this image is also provided as system file. You can
load it with:  system.file("extdata", "L7.MTL.txt", package="water")

Data available from the U.S. Geological Survey.
}
\keyword{datasets}

