% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{LAI}
\alias{LAI}
\title{Estimate LAI from Landsat Data}
\usage{
LAI(method = "metric2010", image, sat = "auto", ESPA = F, aoi, L = 0.1)
}
\arguments{
\item{method}{Method used to estimate LAI from spectral data. See Details.}

\item{image}{image. top-of-atmosphere reflectance for method=="metric" | method=="metric2010" | method=="vineyard" | method=="MCB"; surface reflectance for method = "turner". Not needed if ESPA == TRUE}

\item{sat}{"L7" for Landsat 7, "L8" for Landsat 8 or "auto" to guess from filenames}

\item{ESPA}{Logical. If TRUE will look for espa.usgs.gov related products on working folder}

\item{aoi}{area of interest to crop images, if waterOptions("autoAoi") == TRUE will look for any object called aoi on .GlobalEnv}

\item{L}{L factor used in method = "metric" or "metric2010" to estimate SAVI, defaults to 0.1}
}
\description{
This function implements empirical models to estimate LAI (Leaf Area Index) for satellital images. Models were extracted from METRIC publications and other works developed on different crops.
}
\author{
Guillermo Federico Olmedo

Fonseca-Luengo, David
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

