% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.Leak_Analysis.R
\name{leak_detection}
\alias{leak_detection}
\title{leak_detection}
\usage{
leak_detection(timeseries, method = c("min", "meu"), model_params = NA,
  days = 2, id = NA, pdf_file = NA, out_csv = TRUE,
  return_params = FALSE)
}
\arguments{
\item{timeseries}{a single or a list of xts objects to be investigated for leak detection}

\item{method}{argument for the detection method to be conducted}

\item{model_params}{optional list of model parameters to be used for MEU method, if not provided the model will calculate the parameters and use them for detection}

\item{days}{number of days to be considered for Min method}

\item{id}{optional name of contract id, to be passed in case of conducting detection on single contract}

\item{pdf_file}{name of the pdf file to be used for ouptuing the leak results in a visual form. Defaults to NA, where no pdf will be outputed}

\item{out_csv}{boolean flag indicates whether results should be outputed into csv. Defaults to TRUE}

\item{return_params}{boolean to indicate if the MEU parameters should be returned or not. Defaults to FALSE}
}
\value{
CSV and/or PDF with leak results
}
\description{
This function conducts leak detection on a series or list of series using Min or MEU  method and return results into csv files and/or pdf
}
\examples{
\dontrun{leak_detection( contracts, method= 'meu', pdf = 'Vis', out_csv = TRUE, return_params =FALSE)}
}
