% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leak_Detection_MIN.R
\name{leak_detection_min}
\alias{leak_detection_min}
\title{leak_detection_min}
\usage{
leak_detection_min(input, mID = "000", days = 2)
}
\arguments{
\item{input}{contract timeseries object or list of contracts timeseries}

\item{mID}{contract ID if a single contract is passed to the function}

\item{days}{Number of days to be considered for Analysis}
}
\value{
list of dataframes with data and a booolean flag weather a leak has occured or not
}
\description{
Utility function to highlight leaks based on Min Method, flag as a leak if daily minimum consumption is above 0 for n consequetive days
}
\examples{
\dontrun{leak_detection_min(contract, days =2)}
}
