% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_RSfunctions.R
\name{loadImageSR}
\alias{loadImageSR}
\title{Load Landsat 8 surface reflectance data from folder}
\usage{
loadImageSR(path = getwd(), aoi)
}
\arguments{
\item{path}{folder where band files are stored}

\item{aoi}{area of interest to crop images, if waterOptions("autoAoi") == 
TRUE will look for any object called aoi on .GlobalEnv}
}
\description{
This function loads Landsat bands from a specific folder.
}
\seealso{
Other remote sensing support functions: 
\code{\link{calcRadiance}()},
\code{\link{calcSR}()},
\code{\link{calcTOAr}()},
\code{\link{cfmask}()},
\code{\link{loadImage}()}
}
\author{
Guillermo Federico Olmedo
}
\concept{remote sensing support functions}
