% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_METRIC.R
\name{METRIC.EB}
\alias{METRIC.EB}
\title{Estimates Energy Balance using METRIC2010 Model}
\usage{
METRIC.EB(image.DN, DEM, WeatherStation, aoi, MTL, sat, plain = TRUE,
  thermalband, alb.coeff = "Tasumi", LAI.method = "metric2010",
  Zom.method = "short.crops", anchors.method = "CITRA-MCB",
  ETp.coef = 1.05, Z.om.ws = 0.0018, ESPA = FALSE)
}
\arguments{
\item{image.DN}{}

\item{DEM}{}

\item{WeatherStation}{}

\item{aoi}{}

\item{MTL}{}

\item{sat}{}

\item{plain}{}

\item{thermalband}{}

\item{alb.coeff}{}

\item{LAI.method}{}

\item{Zom.method}{}

\item{anchors.method}{}

\item{ETp.coef}{}

\item{Z.om.ws}{}

\item{ESPA}{}
}
\description{
Estimates Energy Balance using METRIC2010 Model
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

