% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_METRIC.R
\name{METRIC.G}
\alias{METRIC.G}
\title{Estimates Net Radiation as in METRIC Model}
\usage{
METRIC.G(Rn, DEM, image.DN, WeatherStation = WeatherStation)
}
\description{
Estimates Net Radiation as in METRIC Model
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

