% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{METRIC.topo}
\alias{METRIC.topo}
\title{Calculates surface model used in METRIC}
\usage{
METRIC.topo(DEM, result.folder = NULL)
}
\description{
Calculates surface model used in METRIC
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

