% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{momentum.roughness.length}
\alias{momentum.roughness.length}
\title{Calculates Momentum Roughness Length}
\usage{
momentum.roughness.length(method = "short.crops", LAI, NDVI, albedo, a, b,
  fLAI.Perrier, h.Perrier, mountainous = FALSE, surface.model,
  result.folder = NULL)
}
\description{
Calculates Momentum Roughness Length
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

