% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{outgoing.lw.radiation}
\alias{outgoing.lw.radiation}
\title{Calculates Long wave outgoing radiation}
\usage{
outgoing.lw.radiation(LAI, Ts, result.folder = NULL)
}
\description{
Calculates Long wave outgoing radiation
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

