% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{outLWradiation}
\alias{outLWradiation}
\title{Calculates Long wave outgoing radiation}
\usage{
outLWradiation(LAI, Ts)
}
\arguments{
\item{LAI}{raster layer with leaf area index. See LAI()}

\item{Ts}{Land surface temperature. See surfaceTemperature()}
}
\description{
Calculates Long wave outgoing radiation
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

