% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_LSEBclass.R
\name{plot.waterLSEB}
\alias{plot.waterLSEB}
\title{Plot method for waterLSEB S3 class}
\usage{
\method{plot}{waterLSEB}(x, ...)
}
\arguments{
\item{x}{waterLSEB object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot method for waterLSEB S3 class
}
\seealso{
Other LSEB objects related functions: 
\code{\link{print.waterLSEB}()},
\code{\link{writeRaster.waterLSEB}()}
}
\author{
Guillermo Federico Olmedo
}
\concept{LSEB objects related functions}
