% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicsPlot.R
\name{PlotPageFooter}
\alias{PlotPageFooter}
\title{PlotPageFooter}
\usage{
PlotPageFooter()
}
\description{
Finalises the current PDF graphics page by calling dev.off().
Intended to be used after all drawing is complete.
}
