% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicsPlot.R
\name{PlotPageHeader}
\alias{PlotPageHeader}
\title{PlotPageHeader}
\usage{
PlotPageHeader(pi, colT, paper = "A4", orientation = "portrait")
}
\arguments{
\item{pi}{Page index (integer)}

\item{colT}{Colour for text (e.g., colGray)}

\item{paper}{Paper size ("A4" or "A3")}

\item{orientation}{"portrait" or "landscape"}
}
\description{
Draws the page header with the website and page number, adjusting for orientation.
}
