% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{prepareSRTMdata}
\alias{prepareSRTMdata}
\title{Create a mosaic with SRTM grid from image extent}
\usage{
prepareSRTMdata(path = getwd(), format = "tif", extent)
}
\arguments{
\item{format}{format of SRTM grid files}

\item{extent}{minimal extent of mosaic}
}
\description{
Create a mosaic with SRTM grid from image extent
}
\seealso{
Other support functions: \code{\link{checkSRTMgrids}},
  \code{\link{createAoi}}
}
\author{
Guillermo Federico Olmedo
}
