% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_WeatherStation.R
\name{print.waterWeatherStation}
\alias{print.waterWeatherStation}
\title{Print method for waterWeatherStation S3 class}
\usage{
print.waterWeatherStation(WS)
}
\description{
Print method for waterWeatherStation S3 class
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}

