% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_WeatherStation.R
\name{read.WSdata2}
\alias{read.WSdata2}
\title{Prepares weather station data 2}
\usage{
read.WSdata2(WSdata, ..., height = 2.2, lat, long, elev, columns = c("date",
  "time", "radiation", "wind", NA, "RH", "temp", "pp"),
  date.format = "\%d/\%m/\%Y", time.format = "\%H:\%M:\%S",
  datetime.format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "", cf = c(1, 3.6, 1,
  1), MTL)
}
\description{
Prepares weather station data 2
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}

