% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.Utilities.R
\name{read_from_csv}
\alias{read_from_csv}
\title{read_from_csv}
\usage{
read_from_csv(filename, align_and_pad = TRUE,
  align_interval = "30 min", na.action = "None")
}
\arguments{
\item{filename}{The CSV filename containing the three columns as described above}

\item{align_and_pad}{Set to True by default so all the time series' will be aligned to the neareast 30 minutes and all data gaps will be padded with "NA"}

\item{align_interval}{The spacing between reading to be used for alignments}

\item{na.action}{How to handle NAs}
}
\value{
List of XTS objects, each object represents the time series for a specific service agreement
}
\description{
This function receives a csv file with three columns (ID, Time, Litres) which contains meter data
for one or more service agreement. "ID" is a unique identifier for a service agreement, "Time" is
the timestamp for a certain reading in %Y-%m-%d %H:%M:%S format, and "Litres" is the meter reading at
that timestamp. (in units of litres)
}
\examples{
\dontrun{read_from_csv("north_melbourne_data.csv")}
}
