% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{sensible.heat.flux}
\alias{sensible.heat.flux}
\title{Calculates sensible heat flux for METRIC Model}
\usage{
sensible.heat.flux(rho.air, dT, r.ah)
}
\arguments{
\item{rho.air}{- air density}

\item{dT}{- dT from anchor points}

\item{r.ah}{- rugosity}
}
\value{
Returns sensible heat flux
}
\description{
Calculates sensible heat flux for METRIC Model
}
\references{
[1] Allen et al. METRIC Model
}

