\name{set.cores}
\alias{set.cores}
\title{Setting cores}

\usage{set.cores(n = NA)}
\arguments{
  \item{n}{Function sets the number of cores to be used by Rscripts using the parallel package. For older package:multicore use XC's chaos package
The number of cores}
}



\author{Wesley GOI
WesleyGOI}





