% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{soilHeatFlux}
\alias{soilHeatFlux}
\title{Estimates Soil Heat Flux}
\usage{
soilHeatFlux(path = getwd(), image, Ts, albedo, Rn, aoi, sat = "auto",
  ESPA = F)
}
\description{
Estimates Soil Heat Flux
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

