% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{SWtrasmisivity}
\alias{SWtrasmisivity}
\title{Calculates short wave transmisivity}
\usage{
SWtrasmisivity(Kt = 1, ea, dem, incidence.hor)
}
\arguments{
\item{Kt}{unitless turbidity coefficient 0<Kt<=1.0, where Kt=1.0 for clean air and Kt=0.5 for extremely turbid, dusty, or polluted air}

\item{ea}{near-surface vapor pressure (kPa)}

\item{dem}{digital elevation model}

\item{incidence.hor}{solar incidence angle, considering plain surface}
}
\description{
Calculates short wave transmisivity
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

