\name{water-package}
\alias{water-package}
\alias{water}
\docType{package}
\title{water: server}
\description{misc functions}
\details{
\tabular{ll}{Package: \tab water\cr
Maintainer: \tab WesleyGOI <picy2k@gmail.com>\cr
Author: \tab WesleyGOI\cr
Version: \tab 0.0.1\cr
License: \tab GPL-3\cr
Title: \tab water: server\cr}
}
\author{WesleyGOI}

\keyword{ package }


