% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{waterOptions}
\alias{waterOptions}
\title{Global options for water package}
\usage{
waterOptions(format, overwrite, writeResults, destinationFolder, SRTMrepo,
  autoAoi, default = FALSE)
}
\value{
list of the current options (invisibly). If no arguments are provided the options are printed.
}
\description{
This function is based on raster::rasterOptions by Robert Hijmans.
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
Robert J. Hijmans (2015). raster: Geographic Data Analysis and Modeling. R
package version 2.4-18. http://CRAN.R-project.org/package=raster
}

