% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_LSEBclass.R
\name{writeRaster.waterLSEB}
\alias{writeRaster.waterLSEB}
\title{writeRaster method for waterLSEB S3 class}
\usage{
\method{writeRaster}{waterLSEB}(x, ...)
}
\arguments{
\item{x}{waterLSEB object.}

\item{...}{additional parameters to pass to writeRaster()}
}
\description{
writeRaster method for waterLSEB S3 class
}
\seealso{
Other LSEB objects related functions: 
\code{\link{plot.waterLSEB}()},
\code{\link{print.waterLSEB}()}
}
\author{
María Victoria Munafó
}
\concept{LSEB objects related functions}
