% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_sensibleHeatFlux.R
\name{calcH}
\alias{calcH}
\title{Iterative function to estimate H and R.ah}
\usage{
calcH(anchors, Ts, Z.om, WeatherStation, ETp.coef = 1.05, Z.om.ws = 0.0018,
  sat = "auto", ESPA = F, mountainous = FALSE, DEM, Rn, G,
  verbose = FALSE)
}
\arguments{
\item{anchors}{anchors points. Can be the result from calcAnchors() or
a spatialPointDataframe o Dataframe with X, Y, and type. type should be
"cold" or "hot"}

\item{Ts}{Land surface temperature in K. See surfaceTemperature()}

\item{Z.om}{momentum roughness lenght. See momentumRoughnessLength()}

\item{WeatherStation}{WeatherStation data at the flyby from the satellite.
Can be a waterWeatherStation object calculate using read.WSdata and MTL file}

\item{ETp.coef}{ETp coefficient usually 1.05 or 1.2 for alfalfa}

\item{Z.om.ws}{momentum roughness lenght for WeatherStation. Usually
0.0018 or 0.03 for long grass}

\item{sat}{satellite sensor used for NDVI. Can be "L7" or "L8"}

\item{ESPA}{Logical. If TRUE will look for espa.usgs.gov realted
products on working folder}

\item{mountainous}{Logical. If TRUE heat transfer equation will be
adjusted for mountainous terrain}

\item{DEM}{Digital Elevation Model in meters.}

\item{Rn}{Net radiation. See netRadiation()}

\item{G}{Soil Heat Flux. See soilHeatFlux()}

\item{verbose}{Logical. If TRUE will print aditional data to console}
}
\description{
generates an iterative solution to estimate r.ah and H because both are unknown at each pixel.
}
\details{
Sensible heat flux is the rate of heat loss to the air by convection
and conduction, due to a temperature difference.This parameter is computed using the following one-dimensional,
aerodynamic,temperature gradient based equation for heat transport, this method is difficult to solve because
there are two unknowns, rah and dT. To facilitate this computation, METRIC utilize the two "anchor"
pixels  and solve for dT that satisfies eq. given the aerodynamic roughness and wind speed at a given height.
Aerodynamic resistance, and heat transfer is impacted by buoyancy of heated, light air at the surface,
especially when H is large. Therefore, correction to rah is needed to account for buoyancy effects. However,
H is needed to make this correction. An iterative solution for both H and rah is used.
}
\author{
Guillermo Federico Olmedo

de la Fuente-Saiz, Daniel

Fernando Fuentes Penailillo
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007 \cr

Allen, R., Irmak, A., Trezza, R., Hendrickx, J.M.H., Bastiaanssen, W., Kjaersgaard, J., 2011. Satellite-based ET estimation in agriculture using SEBAL and METRIC. Hydrol. Process. 25, 4011-4027. doi:10.1002/hyp.8408 \cr
}

