% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{checkSRTMgrids}
\alias{checkSRTMgrids}
\title{Check needed SRTM grids from image extent}
\usage{
checkSRTMgrids(raw.image)
}
\arguments{
\item{raw.image}{image to calculate extent}
}
\description{
Check needed SRTM grids from image extent
}
\author{
Guillermo Federico Olmedo
}

