% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEM_Talca}
\alias{DEM_Talca}
\title{SRTM DEM from east of Talca, Chile}
\format{
An object of class \code{RasterLayer} of dimension 417 x 508 x 1.
}
\source{
\url{https://www.usgs.gov/}
}
\usage{
DEM_Talca
}
\description{
A RasterLayer object with a a Digital Elevation Model.
}
\details{
Original data is float. Example data is integer.

Data available from the U.S. Geological Survey.
}
\keyword{datasets}
