% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_WeatherStation.R
\name{getDataWS}
\alias{getDataWS}
\title{Export data.frame from waterWeatherStation Object}
\usage{
getDataWS(WeatherStation)
}
\arguments{
\item{WeatherStation}{waterWeatherStation object. See read.WSdata()}
}
\description{
Export weather conditions at satellite flyby from waterWeatherStation object
}
\author{
Guillermo Federico Olmedo
}
