% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{incSWradiation}
\alias{incSWradiation}
\title{Calculates Incoming Solar Radiation}
\usage{
incSWradiation(surface.model, solar.angles, WeatherStation, ...)
}
\arguments{
\item{surface.model}{surface.model with dem aspect and slope}

\item{solar.angles}{...}

\item{WeatherStation}{...}
}
\description{
This function calculates incoming solar radiation from surface model and solar angles.
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}

Daniel de la Fuente Saiz
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

