% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{incSWradiation}
\alias{incSWradiation}
\title{Calculates Incoming Solar Radiation}
\usage{
incSWradiation(surface.model, solar.angles, WeatherStation)
}
\arguments{
\item{surface.model}{rasterStack with DEM, Slope and Aspect. See surface.model()}

\item{solar.angles}{rasterStack with latitude, declination, hour.angle, incidence.hor and incidence.rel. See solarAngles()}

\item{WeatherStation}{Weather Station data}
}
\description{
This function calculates incoming solar radiation from surface model and solar angles.
}
\author{
Guillermo Federico Olmedo

Daniel de la Fuente Saiz

Fonseca-Luengo, David
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007 \cr
}

