% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L7_Talca}
\alias{L7_Talca}
\title{Landsat 7 scene from east of Talca, Chile}
\format{
An object of class \code{RasterBrick} of dimension 417 x 508 x 7.
}
\source{
\url{https://www.usgs.gov/}
}
\usage{
L7_Talca
}
\description{
A RasterStack object with a Landsat 7 ETM+ subset image. 
Band names are stored in layer names. Images are USGS L1 raw data.
}
\details{
Metadata for this image is also provided as system file. You can
load it with:  system.file("extdata", "L7.MTL.txt", package="water")

Data available from the U.S. Geological Survey.
}
\keyword{datasets}
