% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{LAI}
\alias{LAI}
\title{Estimate LAI from Landsat Data}
\usage{
LAI(method = "metric2010", path = getwd(), aoi, L = 0.1, ESPA = F,
  image, sat = "auto")
}
\description{
Estimate LAI from Landsat Data
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

