% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_METRIC.R
\name{METRIC.EB}
\alias{METRIC.EB}
\title{Estimates Energy Balance using METRIC2010 Model}
\usage{
METRIC.EB(path = getwd(), image.DN, DEM, WeatherStation, aoi, plain = TRUE)
}
\description{
Estimates Energy Balance using METRIC2010 Model
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

