% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_METRIC.R
\name{METRIC.EB}
\alias{METRIC.EB}
\title{Estimates Energy Balance using METRIC2010 Model}
\usage{
METRIC.EB(
  image.DN,
  image.SR,
  WeatherStation,
  MTL,
  sat = "auto",
  thermalband,
  plain = TRUE,
  DEM,
  aoi,
  G.method = "Tasumi",
  alb.coeff = "Tasumi",
  LST.method = "SC",
  LAI.method = "metric2010",
  L = 0.1,
  Zom.method = "short.crops",
  anchors.method = "CITRA-MCB",
  anchors,
  n = 1,
  ETp.coef = 1.05,
  Z.om.ws = 0.0018,
  verbose = FALSE,
  extraParameters = vector()
)
}
\arguments{
\item{image.DN}{raw imagen in digital counts to evaluate}

\item{image.SR}{L8 ONLY. Surface reflectance imagen. water package does not 
include a model to calculate surface reflectance for Landsat 8 images. Landsat 8 users 
should download precalculated surface reflectances from espa website 
(espa.cr.usgs.gov).}

\item{WeatherStation}{Weather Station data, can be a waterWeatherStation 
object}

\item{MTL}{Landsat metadata file}

\item{sat}{Landsat satellite version. "L7" or "L8"}

\item{thermalband}{Landsat low gain thermalband}

\item{plain}{Logical. If TRUE surface is assumed plain}

\item{DEM}{Digital Elevation Model of the study area. Not needed
if plain = TRUE}

\item{aoi}{SpatialPolygon object with limits of Area of interest}

\item{G.method}{method used for the G estimation. Currently implemeted are 
"Tasumi" for Tasumi,2003 or "Bastiaanssen" for Bastiaanssen, 2000}

\item{alb.coeff}{coefficient to transform narrow to broad band albedo.
See Details.}

\item{LST.method}{Method for land surface temperature estimation. "SC" 
for single channel or "SW" for split window algorithm. 
"SW" is only available for L8. See \code{water::surfaceTemperature}}

\item{LAI.method}{Method used to estimate LAI from spectral data. 
See Details.}

\item{L}{L value for SAVI calculation}

\item{Zom.method}{method selected to calculate momentum roughness 
length. Use "short.crops" for short crops methods from Allen et al (2007); 
"custom" for custom method also in Allen et al (2007); Or "Perrier" to use 
Perrier equation as in Santos et al (2012) and Pocas et al (2014).}

\item{anchors.method}{method for the automatic selection of the anchor pixels.}

\item{anchors}{data.frame or SpatialPointsDataFrame with the anchor
pixels. The data frame must include a "type" column 
with "hot" and "cold" values.}

\item{n}{number of pair of anchors pixels to calculate}

\item{ETp.coef}{ETp coefficient usually 1.05 or 1.2 for alfalfa}

\item{Z.om.ws}{momentum roughness lenght for WeatherStation. Usually
0.0018 or 0.03 for long grass}

\item{verbose}{Logical. If TRUE will print aditional data to console}

\item{extraParameters}{Extra parameters for the non default methods. i.e. 
Zom.method = "Perrier", needs two extra parameters: fLAI, h. See 
help(momentumRoughnessLength).}
}
\description{
Estimates Energy Balance using METRIC2010 Model
}
\details{
There are differents models to convert narrowband data to broadband albedo. 
You can choose alb.coeff ="Tasumi" to use Tasumi et al (2008) coefficients, 
calculated for Landsat 7; alb.coeff ="Liang" to use Liang Landsat 7 
coefficients or "Olmedo" to use Olmedo coefficients for Landsat 8.
}
\section{Extra Parameters}{

Extra Paramenters for functions inside METRIC.EB() include:
* for momentumRoughness when Zom.method = "Perrier": fLAI, h.
* for calcAnchors(): minDist, WSbuffer, deltaTemp
}

\examples{
### Data preparation
library(water)
aoi <- createAoi(topleft = c(500000, -3644000), bottomright = c(526000, -3660000))
raw_data_folder <- system.file("extdata", package="water")
image <- loadImage(path=raw_data_folder, aoi=aoi, sat="L8")
image.SR <- loadImageSR(path=raw_data_folder, aoi=aoi)
csvfile <- system.file("extdata", "INTA.csv", package="water")
MTLfile <- system.file("extdata", "LC82320832016040LGN00_MTL.txt", package="water")
\dontrun{
WeatherStation <- read.WSdata(WSdata = csvfile, 
                              datetime.format =  "\%Y/\%m/\%d \%H:\%M", 
                              columns = c("datetime", "temp",
                                          "RH", "pp", "radiation", "wind"), 
                              lat=-33.00513, long= -68.86469, elev=927, height= 2,
                              MTL=MTLfile)
                              
### LSEB with default methods and no extra parameters                               
Energy.Balance <- METRIC.EB(image.DN = image, image.SR = image.SR,
                            plain=TRUE, aoi=aoi, n = 5, WeatherStation = WeatherStation, 
                            ETp.coef = 1.2, sat="L8", alb.coeff = "Olmedo", LST.method = "SW", 
                            LAI.method = "metric2010", Z.om.ws = 0.03, MTL = MTLfile)
                            
                            
### LSEB with "Perrier" method for Zom and extra parameters                               
Energy.Balance <- METRIC.EB(image.DN = image, image.SR = image.SR,
                            plain=TRUE, aoi=aoi, n = 5, WeatherStation = WeatherStation, 
                            ETp.coef = 1.2, sat="L8", alb.coeff = "Olmedo", LST.method = "SW", 
                            LAI.method = "metric2010", Zom.method = "Perrier", Z.om.ws = 0.03, 
                            MTL = MTLfile, extraParameters = c(fLAI = 0.5, h = 1.8) ) 
}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
González, Arturo & Hay, Christopher & Kjaersgaard, Jeppe & Neale, Christopher. (2015). Use of Remote Sensing to Generate Crop Coefficient and Estimate Actual Crop Evapotranspiration. 10.13031/aim.20152190105.
}
\seealso{
Other METRIC model functions: 
\code{\link{METRIC.G}()},
\code{\link{METRIC.Rn}()}
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\concept{METRIC model functions}
