% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_METRIC.R
\name{METRIC.G}
\alias{METRIC.G}
\title{Estimates Net Radiation as in METRIC Model}
\usage{
METRIC.G(image.DN, WeatherStation = WeatherStation, Rn, plain = TRUE, DEM, aoi)
}
\arguments{
\item{image.DN}{raw imagen in digital counts to evaluate}

\item{WeatherStation}{Weather Station data, can be a waterWeatherStation 
object}

\item{Rn}{RasterLayer with Net Radiation data in W/m2}

\item{plain}{Logical. If TRUE surface is assumed plain}

\item{DEM}{Digital Elevation Model of the study area. Not needed
if plain = TRUE}

\item{aoi}{SpatialPolygon object with limits of Area of interest}
}
\description{
Estimates Net Radiation as in METRIC Model
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}
\seealso{
Other METRIC model functions: 
\code{\link{METRIC.EB}()},
\code{\link{METRIC.Rn}()}
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\concept{METRIC model functions}
