% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{METRICtopo}
\alias{METRICtopo}
\title{Calculates surface model used in METRIC}
\usage{
METRICtopo(DEM)
}
\arguments{
\item{DEM}{raster with Digital elevation model}
}
\description{
DEM map is used to generate the surface representation of the image through of aspect and slope maps. This procedure helps to avoid differences in the surface temperature (and finally Evapotranspiration) caused by different incidence angles and/or elevations in mountainous areas.
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007 \cr
}
\seealso{
Other net radiation related functions: 
\code{\link{LAI}()},
\code{\link{SWtrasmisivity}()},
\code{\link{albedo}()},
\code{\link{incLWradiation}()},
\code{\link{incSWradiation}()},
\code{\link{netRadiation}()},
\code{\link{outLWradiation}()},
\code{\link{soilHeatFlux}()},
\code{\link{solarAngles}()},
\code{\link{surfaceTemperature}()}
}
\author{
Guillermo Federico Olmedo

Fonseca-Luengo, David
}
\concept{net radiation related functions}
