% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{netRadiation}
\alias{netRadiation}
\title{Estimates net radiation}
\usage{
netRadiation(LAI, albedo, Rs.inc, Rl.inc, Rl.out)
}
\arguments{
\item{LAI}{raster layer with leaf area index. See LAI()}

\item{albedo}{broadband surface albedo. See albedo()}

\item{Rs.inc}{incoming short-wave radiation}

\item{Rl.inc}{incomin long-wave radiation}

\item{Rl.out}{outgoing long-wave radiation}
}
\description{
Estimates net radiation
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

