% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{netRadiation}
\alias{netRadiation}
\title{Estimates net radiation}
\usage{
netRadiation(LAI, albedo, Rs.inc, Rl.inc, Rl.out)
}
\arguments{
\item{LAI}{raster layer with leaf area index. See LAI()}

\item{albedo}{broadband surface albedo. See albedo()}

\item{Rs.inc}{incoming short-wave radiation}

\item{Rl.inc}{incomin long-wave radiation}

\item{Rl.out}{outgoing long-wave radiation}
}
\description{
This function estimates net radiation considering a surface radiation balance. Equations use the information from the image, in addition of measurements of actual vapor pressure and altitude.
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007 \cr
}
\seealso{
Other net radiation related functions: 
\code{\link{LAI}()},
\code{\link{METRICtopo}()},
\code{\link{SWtrasmisivity}()},
\code{\link{albedo}()},
\code{\link{incLWradiation}()},
\code{\link{incSWradiation}()},
\code{\link{outLWradiation}()},
\code{\link{soilHeatFlux}()},
\code{\link{solarAngles}()},
\code{\link{surfaceTemperature}()}
}
\author{
Guillermo Federico Olmedo

Fonseca-Luengo, David
}
\concept{net radiation related functions}
