% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{outLWradiation}
\alias{outLWradiation}
\title{Calculates Long wave outgoing radiation}
\usage{
outLWradiation(LAI, Ts)
}
\arguments{
\item{LAI}{raster layer with leaf area index. See LAI()}

\item{Ts}{Land surface temperature. See surfaceTemperature()}
}
\description{
This function estimates the long wave outgoing radiation using the Stefan-Boltzmann equation.
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for 
mapping evapotranspiration with internalized calibration (METRIC) - Model" 
Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007 \cr
}
\seealso{
Other net radiation related functions: 
\code{\link{LAI}()},
\code{\link{METRICtopo}()},
\code{\link{SWtrasmisivity}()},
\code{\link{albedo}()},
\code{\link{incLWradiation}()},
\code{\link{incSWradiation}()},
\code{\link{netRadiation}()},
\code{\link{soilHeatFlux}()},
\code{\link{solarAngles}()},
\code{\link{surfaceTemperature}()}
}
\author{
Guillermo Federico Olmedo

Fonseca-Luengo, David
}
\concept{net radiation related functions}
