% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_WeatherStation.R
\name{plot.waterWeatherStation}
\alias{plot.waterWeatherStation}
\title{Plot method for waterWeatherStation S3 class}
\usage{
plot.waterWeatherStation(WS, alldata = TRUE)
}
\description{
Plot method for waterWeatherStation S3 class
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}

