% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_WeatherStation.R
\name{plot.waterWeatherStation}
\alias{plot.waterWeatherStation}
\title{Plot method for waterWeatherStation S3 class}
\usage{
\method{plot}{waterWeatherStation}(x, alldata = TRUE, ...)
}
\arguments{
\item{x}{waterWeatherStation object. See read.WSdata()}

\item{alldata}{plot all data on waterWeatherStation object. If false, will
only plot hourly data.}

\item{...}{additional parameters to pass to plot()}
}
\description{
Plot method for waterWeatherStation S3 class
}
\author{
Guillermo Federico Olmedo
}

