% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_WeatherStation.R
\name{plot.waterWeatherStation}
\alias{plot.waterWeatherStation}
\title{Plot method for waterWeatherStation S3 class}
\usage{
\method{plot}{waterWeatherStation}(x, hourly = FALSE, sat = TRUE, date, ...)
}
\arguments{
\item{x}{waterWeatherStation object. See read.WSdata()}

\item{hourly}{If TRUE will plot only hourly data, instead of all data 
from the waterWeatherStation object}

\item{sat}{If TRUE, and if the waterWeatherStation object was created
using a Landsat Metadata File, will plot only data from day of the satellite
overpass}

\item{date}{When sat = FALSE, the date to plot in a multi-day object}

\item{...}{additional parameters to pass to plot()}
}
\description{
Plot method for waterWeatherStation S3 class
}
\seealso{
Other Weather station related functions: 
\code{\link{print.waterWeatherStation}()},
\code{\link{read.WSdata2}()},
\code{\link{read.WSdata}()}
}
\author{
Guillermo Federico Olmedo
}
\concept{Weather station related functions}
