% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicsPlot.R
\name{PlotBasePage}
\alias{PlotBasePage}
\title{PlotBasePage}
\usage{
PlotBasePage(
  path = "",
  pagesize = "A4",
  orientation = "portrait",
  DEBUG = FALSE
)
}
\arguments{
\item{path}{File path for output PDF}

\item{pagesize}{Paper size ("A4" or "A3")}

\item{orientation}{"portrait" or "landscape"}

\item{DEBUG}{Logical; draw axes if TRUE and label page format}
}
\description{
Sets up a blank PDF canvas with given page size and orientation.
}
