% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{prepareSRTMdata}
\alias{prepareSRTMdata}
\title{Create a mosaic with SRTM grid from image extent}
\usage{
prepareSRTMdata(path = getwd(), format = "tif", extent = image.DN,
  result.folder = NULL)
}
\description{
Create a mosaic with SRTM grid from image extent
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}

