% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{prepareSRTMdata}
\alias{prepareSRTMdata}
\title{Create a mosaic with SRTM grid from image extent}
\usage{
prepareSRTMdata(path = getwd(), format = "tif", extent = image.DN)
}
\description{
Create a mosaic with SRTM grid from image extent
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}

