% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{prepareSRTMdata}
\alias{prepareSRTMdata}
\title{Create a mosaic with SRTM grid from image extent}
\usage{
prepareSRTMdata(format = "tif", extent)
}
\arguments{
\item{format}{format of SRTM grid files}

\item{extent}{minimal extent of mosaic}
}
\description{
Create a mosaic with SRTM grid from image extent
}
\author{
Guillermo Federico Olmedo
}
