% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_WeatherStation.R
\name{print.waterWeatherStation}
\alias{print.waterWeatherStation}
\title{Print method for waterWeatherStation S3 class}
\usage{
\method{print}{waterWeatherStation}(x, ...)
}
\arguments{
\item{x}{waterWeatherStation object. See read.WSdata()}

\item{...}{additional parameters to pass to print()}
}
\description{
Print method for waterWeatherStation S3 class
}
\seealso{
Other Weather station related functions: 
\code{\link{plot.waterWeatherStation}()},
\code{\link{read.WSdata2}()},
\code{\link{read.WSdata}()}
}
\author{
Guillermo Federico Olmedo
}
\concept{Weather station related functions}
