% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water_ET.R
\name{read.WSdata}
\alias{read.WSdata}
\title{Prepares weather station data}
\usage{
read.WSdata(WSdata, ..., height = 2.2, lat, long, elev, columns = c("date",
  "time", "radiation", "wind", NA, "RH", "temp", "pp"),
  date.format = "\%Y-\%m-\%d", time.format = "\%H:\%M:\%S",
  datetime.format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "", cf = c(1, 1, 1,
  1), MTL)
}
\description{
Prepares weather station data
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}

