% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/L8MfR_functions.R
\name{sensible.heat.flux}
\alias{sensible.heat.flux}
\title{Calculates sensible heat flux for METRIC Model}
\usage{
sensible.heat.flux(rho.air, dT, r.ah)
}
\arguments{
\item{rho.air}{- air density}

\item{dT}{- dT from anchor points}

\item{r.ah}{- rugosity}
}
\value{
Returns sensible heat flux
}
\description{
Calculates sensible heat flux for METRIC Model
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for mapping evapotranspiration with internalized calibration (METRIC) - Model" Journal of Irrigation and Drainage Engineering, vol. 133, p. 380, 2007
}

