% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_netRadiation.G.R
\name{surfaceTemperature}
\alias{surfaceTemperature}
\title{Estimates Land Surface Temperature from Landsat Data}
\usage{
surfaceTemperature(
  image.DN,
  sat = "auto",
  LAI,
  aoi,
  method = "SC",
  WeatherStation,
  thermalband
)
}
\arguments{
\item{image.DN}{raw image in digital numbers}

\item{sat}{"L7" for Landsat 7, "L8" for Landsat 8 or "auto" to guess from filenames}

\item{LAI}{raster layer with leaf area index. See LAI()}

\item{aoi}{area of interest to crop images, if 
waterOptions("autoAoi") == TRUE will look for any 
object called aoi on .GlobalEnv}

\item{method}{"SC" for single channel or "SW" for split window 
algorithm. "SW" is only available for L8}

\item{WeatherStation}{Weather Station data}

\item{thermalband}{Satellite thermal band. For L8 this should be band 10. Deprecated argument}
}
\description{
Surface temperature is estimated using a modified Plank equation considering empirical constants for Landsat images. In addition, this model implements a correction of thermal radiance following to Wukelic et al. (1989).
}
\references{
R. G. Allen, M. Tasumi, and R. Trezza, "Satellite-based energy balance for 
mapping evapotranspiration with internalized calibration (METRIC) - 
Model" Journal of Irrigation and Drainage Engineering, vol. 133, 
p. 380, 2007 \cr

Wukelic G. E.; Gibbons D. E.; Martucci L. M. & Foote, H. P. Radiometric 
calibration of Landsat thematic mapper thermal band Remote Sensing of 
Environment, 1989, 28, (339-347) \cr

Jimenez-Munoz, J. C., Sobrino, J. A., Skokovic, D., Mattar, C., & Cristobal, 
J. (2014). Land surface temperature retrieval methods from landsat-8 thermal 
infrared sensor data.  IEEE Geoscience and Remote Sensing Letters, 11(10), 
1840–1843. http://doi.org/10.1109/LGRS.2014.2312032 \cr
}
\seealso{
Other net radiation related functions: 
\code{\link{LAI}()},
\code{\link{METRICtopo}()},
\code{\link{SWtrasmisivity}()},
\code{\link{albedo}()},
\code{\link{incLWradiation}()},
\code{\link{incSWradiation}()},
\code{\link{netRadiation}()},
\code{\link{outLWradiation}()},
\code{\link{soilHeatFlux}()},
\code{\link{solarAngles}()}
}
\author{
Guillermo Federico Olmedo

Fonseca-Luengo, David
}
\concept{net radiation related functions}
