% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{waterOptions}
\alias{waterOptions}
\title{Global options for water package}
\usage{
waterOptions(
  overwrite,
  writeResults,
  outputFolder,
  SRTMrepo,
  autoAoi,
  default = FALSE
)
}
\arguments{
\item{overwrite}{Logical. If TRUE and writeResults is TRUE it will 
overwrite results. If FALSE, results are save with a name with name_datetime.}

\item{writeResults}{Logical. If TRUE it'll write result to disk. This is 
slower but if FALSE you can have out-of-memory problems.}

\item{outputFolder}{Name of a folder to save files, relative to workind
folder.}

\item{SRTMrepo}{A folder where SRTM grids are stored, to create DEM. See
prepareSRTMdata()}

\item{autoAoi}{Logical. If TRUE it'll look for a object called aoi on 
.GlobalEnv and use it as aoi. See createAoi()}

\item{default}{Logical. If TRUE will revert all options to defaults 
values}
}
\value{
list of the current options (invisibly). If no arguments are provided the options are printed.
}
\description{
This function is based on raster::rasterOptions by Robert Hijmans.
}
\references{
Robert J. Hijmans (2015). raster: Geographic Data Analysis and Modeling. R
package version 2.4-18. http://CRAN.R-project.org/package=raster
}
\author{
Guillermo F Olmedo, \email{guillermo.olmedo@gmail.com}
}
